fix: refactored connection into own rust file, and created publisher mod.
