+++ alicemq/src/publisher.rs
@@ -1,0 +1,4 @@
+
+pub struct Publisher {
+
+}
